refactored database code and added transfer model
